# ios/SplitBackAPI/Sources/SplitBackAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SplitBack",
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/splitwise/login": {
      "get": {
        "tags": [
          "splitwise-auth"
        ],
        "summary": "Login",
        "operationId": "login_auth_splitwise_login_get",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "matt",
              "title": "User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/splitwise/callback": {
      "get": {
        "tags": [
          "splitwise-auth"
        ],
        "summary": "Callback",
        "operationId": "callback_auth_splitwise_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/me": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Me",
        "operationId": "me_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "List Users",
        "operationId": "list_users_users_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/UserSource",
              "title": "Source"
            }
          },
          {
            "name": "updated_since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "Updated Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  },
                  "title": "Response List Users Users Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Create User",
        "operationId": "create_user_users_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": [
          "users"
        ],
        "summary": "Get User",
        "operationId": "get_user_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users"
        ],
        "summary": "Update User",
        "operationId": "update_user_users__user_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "summary": "Delete User",
        "operationId": "delete_user_users__user_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/splitwise/status": {
      "get": {
        "tags": [
          "splitwise"
        ],
        "summary": "Status",
        "operationId": "status_splitwise_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitwiseStatus"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/splitwise/import": {
      "post": {
        "tags": [
          "splitwise"
        ],
        "summary": "Run Import",
        "operationId": "run_import_splitwise_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitwiseImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SplitwiseImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/splitwise/groups/{group_id}/import-local": {
      "post": {
        "tags": [
          "splitwise"
        ],
        "summary": "Import Group Local",
        "description": "Clone a Splitwise-linked group into a new self-hosted group (native, full-featured\ncopies), then archive the source so balances don't double-count. Archived source\nexpenses are not copied; receipts/line items copy if present.",
        "operationId": "import_group_local_splitwise_groups__group_id__import_local_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Group Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocalImportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BackendType": {
        "type": "string",
        "enum": [
          "self_hosted",
          "splitwise"
        ],
        "title": "BackendType"
      },
...
      "GroupResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "backend_type": {
            "$ref": "#/components/schemas/BackendType"
          },
          "splitwise_group_id": {
            "type": "string",
            "title": "Splitwise Group Id"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden"
          },
          "archived_at": {
            "type": "string",
            "format": "date-time",
            "title": "Archived At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "backend_type",
          "hidden",
          "created_at",
          "updated_at"
        ],
        "title": "GroupResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalImportRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "title": "LocalImportRequest"
      },
      "LocalImportResult": {
        "properties": {
          "group": {
            "$ref": "#/components/schemas/GroupResponse"
          },
          "expenses_copied": {
            "type": "integer",
            "title": "Expenses Copied"
          }
        },
        "type": "object",
        "required": [
          "group",
          "expenses_copied"
        ],
        "title": "LocalImportResult"
      },
      "MeResponse": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "authenticated": {
            "type": "boolean",
            "title": "Authenticated"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "MeResponse"
      },
...
      "SplitwiseImportRequest": {
        "properties": {
          "since": {
            "type": "string",
            "title": "Since"
          },
          "until": {
            "type": "string",
            "title": "Until"
          },
          "as_user": {
            "type": "string",
            "title": "As User"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "default": false
          }
        },
        "type": "object",
        "title": "SplitwiseImportRequest"
      },
      "SplitwiseImportResult": {
        "properties": {
          "groups": {
            "type": "integer",
            "title": "Groups"
          },
          "expenses_fetched": {
            "type": "integer",
            "title": "Expenses Fetched"
          },
          "imported": {
            "type": "integer",
            "title": "Imported"
          },
          "skipped_deleted": {
            "type": "integer",
            "title": "Skipped Deleted"
          },
          "settle_ups": {
            "type": "integer",
            "title": "Settle Ups"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "users": {
            "type": "integer",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "expenses_fetched",
          "imported",
          "skipped_deleted",
          "settle_ups",
          "dry_run"
        ],
        "title": "SplitwiseImportResult"
      },
      "SplitwiseStatus": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          }
        },
        "type": "object",
        "required": [
          "connected",
          "users"
        ],
        "title": "SplitwiseStatus"
      },
...
      "UserCreate": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "source": {
            "$ref": "#/components/schemas/UserSource",
            "default": "manual"
          },
          "splitwise_user_id": {
            "type": "string",
            "title": "Splitwise User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "UserCreate"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "source": {
            "$ref": "#/components/schemas/UserSource"
          },
          "splitwise_user_id": {
            "type": "string",
            "title": "Splitwise User Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "avatar_url": {
            "type": "string",
            "title": "Avatar Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "identifier",
          "display_name",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse"
      },
      "UserSource": {
        "type": "string",
        "enum": [
          "app",
          "manual",
          "splitwise"
        ],
        "title": "UserSource"
      },
      "UserUpdate": {
        "properties": {
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          }
        },
        "type": "object",
        "title": "UserUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}